+++ src/components/Editor/Menu/Image/DropShadow.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { changeOperationData, getHeight, getOperationData, getWidth } from '../../../store/editorSlice';
+import { changeOperationData, getHeight, getOperationData, getWidth } from '../../../../store/editorSlice';
 import { ColorResult, CustomPicker, HuePicker } from 'react-color';
-import NumberInput from './NumberInput';
-import Layer from '../../../pixit/Layer';
-import menuSection from './MenuSection';
-import './menu-section.css';
+import NumberInput from '../common/NumberInput';
+import Layer from '../../../../pixit/Layer';
+import menuSection from '../common/MenuSection';
+import '../menu-section.css';
 
 function ShadowOffsetSection() {
   const operationData = useSelector(getOperationData);
