+++ src/components/Editor/Menu/Image/ResizeCanvas.tsx
@@ -1,0 +1,4 @@
+import menuSection from '../common/MenuSection';
+import PixelSection from '../common/PixelSection';
+
+export default menuSection('Pixels', PixelSection);
