+++ src/components/Editor/Menu/Image/RotateImage.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { changeOperationData, getHeight, getOperationData, getWidth } from '../../../store/editorSlice';
-import menuSection from './MenuSection';
-import NumberInput from './NumberInput';
-import './menu-section.css';
+import { changeOperationData, getHeight, getOperationData, getWidth } from '../../../../store/editorSlice';
+import menuSection from '../common/MenuSection';
+import NumberInput from '../common/NumberInput';
+import '../menu-section.css';
 
 function PivotSection() {
   const canvasWidth = useSelector(getWidth);
