--- src/components/Editor/Menu/ResizeCanvas.tsx
@@ -1,4 +1,0 @@
-import menuSection from './MenuSection';
-import PixelSection from './PixelSection';
-
-export default menuSection('Pixels', PixelSection);
